backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
deepgram-sdk==3.2.0
groq==0.4.1
aiofiles==23.2.1